Replace typescript files with type declarations
Remove typescript builds
