+++ index.js
@@ -1,0 +1,13 @@
+const {authenticatedLndGrpc} = require('./lnd_grpc');
+const {emitGrpcEvents} = require('./lnd_gateway');
+const {grpcRouter} = require('./lnd_gateway');
+const {lndGateway} = require('./lnd_gateway');
+const {unauthenticatedLndGrpc} = require('./lnd_grpc');
+
+module.exports = {
+  authenticatedLndGrpc,
+  emitGrpcEvents,
+  grpcRouter,
+  lndGateway,
+  unauthenticatedLndGrpc,
+};
