+++ package.json
@@ -47,13 +47,8 @@
     "url": "https://github.com/alexbosworth/lightning.git"
   },
   "scripts": {
-    "prepare": "rm -rf dist && npm run tsc",
-    "test": "tap test/arrays/*.js test/bolt00/*.js test/bolt02/*.js test/lnd_gateway/*.js test/lnd_grpc/*.js test/lnd_methods/address/*.js test/lnd_methods/generic/*.js test/lnd_methods/info/*.js test/lnd_methods/invoices/*.js test/lnd_methods/macaroon/*.js test/lnd_methods/offchain/*.js test/lnd_methods/onchain/*.js test/lnd_methods/peers/*.js test/lnd_requests/*.js test/lnd_responses/*.js",
-    "tsc": "./node_modules/typescript/bin/tsc --build tsconfig.json"
+    "test": "tap test/arrays/*.js test/bolt00/*.js test/bolt02/*.js test/lnd_gateway/*.js test/lnd_grpc/*.js test/lnd_methods/address/*.js test/lnd_methods/generic/*.js test/lnd_methods/info/*.js test/lnd_methods/invoices/*.js test/lnd_methods/macaroon/*.js test/lnd_methods/offchain/*.js test/lnd_methods/onchain/*.js test/lnd_methods/peers/*.js test/lnd_requests/*.js test/lnd_responses/*.js"
   },
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
   "version": "3.0.15"
 }
