--- tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2017",
-    "sourceMap": true,
-    "lib": [
-      "es2017"
-    ],
-    "declaration": true,
-    "outDir": "dist"
-  },
-  "compileOnSave": true,
-  "exclude": [
-    "node_modules",
-    "dist"
-  ]
-}
