+++ typescript/lnd_gateway.d.ts
@@ -1,0 +1,37 @@
+import type * as request from 'request';
+import WebSocket = require('ws');
+import {AuthenticatedLnd} from './authenticated_lnd_grpc';
+import {UnauthenticatedLnd} from './unauthenticated_lnd_grpc';
+
+export type UnauthenticatedLndGatewayServer = {
+  /** Base64 or Hex Serialized Gateway TLS Cert String */
+  cert?: string;
+  /** Request Function */
+  request: (
+    options: request.Options,
+    callback: request.RequestCallback
+  ) => request.Request;
+  /** LND Gateway URL String */
+  url: string;
+  /** Websocket Constructor Function */
+  websocket: typeof WebSocket;
+};
+
+export type AuthenticatedLndGatewayServer = UnauthenticatedLndGatewayServer & {
+  /** Use Base 64 Encoded Macaroon String */
+  macaroon: string;
+};
+
+/**
+ * Interface to an LND gateway server.
+ */
+export function lndGateway(
+  server: AuthenticatedLndGatewayServer
+): {lnd: AuthenticatedLnd};
+
+/**
+ * Interface to an LND gateway server.
+ */
+export function lndGateway(
+  server: UnauthenticatedLndGatewayServer
+): {lnd: UnauthenticatedLnd};
